Split up ML allowlist in a separate file
